+++ .idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/CSCB532-logistics-company.iml" filepath="$PROJECT_DIR$/.idea/CSCB532-logistics-company.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/logisticsCompany.main.iml" filepath="$PROJECT_DIR$/.idea/modules/logisticsCompany.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/logisticsCompany.test.iml" filepath="$PROJECT_DIR$/.idea/modules/logisticsCompany.test.iml" />
     </modules>
   </component>
 </project>
