+++ src/main/java/com/logistics/logisticsCompany/LogisticsCompanyApplication.java
@@ -1,11 +1,7 @@
-package com.logistics.logisticsCompany.organizaiton;
+package com.logistics.logisticsCompany;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 
 @SpringBootApplication
