+++ src/main/resources/application.yml
@@ -1,2 +1,2 @@
 server:
-  port: 8081
+  port: 8082
